a few updates (stop_on_error, do_dedent, out of the list index, oneline_change)(#6)
use stop_on_error instead of ignore_error
deal with the out of bound issue for the list index
default do_dedent to False
no need to pass the "oneline_change" parameter when calling modify(...), determine it automatically

+++ tests/test.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+module load python
+if ! which python &>/dev/null; then
+  echo "python not found"
+  exit 1
+fi
 
 rm -rf tmp
 mkdir -p tmp
